# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
  HTTP500InternalServerError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
info:
  contact:
    email: support@ablecloud.io
    name: 윤여천
    url: http://www.ablecloud.io
  description: This is a GlueAPI server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Glue-API
  version: v1.0
paths:
...
  /api/v1/rgw:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: RADOS Gateway Daemon 정보를 가져옵니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                default:
                  type: boolean
                id:
                  type: string
                port:
                  type: integer
                realm_name:
                  type: string
                server_hostname:
                  type: string
                service_map_id:
                  type: string
                version:
                  type: string
                zone_name:
                  type: string
                zonegroup_name:
                  type: string
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show List of RADOS Gateway Daemon
      tags:
      - RGW
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: RADOS Gateway Service를 생성합니다.
      parameters:
      - description: RGW Service Name
        in: formData
        name: service_name
        required: true
        type: string
      - description: RGW Realm Name
        in: formData
        name: realm_name
        type: string
      - description: RGW Zone Group Name
        in: formData
        name: zonegroup_name
        type: string
      - description: RGW Zone Name
        in: formData
        name: zone_name
        type: string
      - description: 'Service Port(default: 80)'
        in: formData
        name: port
        type: integer
      - collectionFormat: multi
        description: Service Placement Host Name
        in: formData
        items:
          type: string
        name: hostname
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Create of RADOS Gateway Service
      tags:
      - RGW
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: RADOS Gateway Service를 수정합니다.
      parameters:
      - description: RGW Service Name
        in: formData
        name: service_id
        required: true
        type: string
      - description: RGW Realm Name
        in: formData
        name: realm_name
        type: string
      - description: RGW Zone Group Name
        in: formData
        name: zonegroup_name
        type: string
      - description: RGW Zone Name
        in: formData
        name: zone_name
        type: string
      - description: 'Service Port(default: 80)'
        in: formData
        name: port
        type: integer
      - collectionFormat: multi
        description: Service Placement Hosts
        in: formData
        items:
          type: string
        name: hosts
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Update of RADOS Gateway Service
      tags:
      - RGW
  /api/v1/rgw/bucket:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      description: RADOS Gateway 버킷을 삭제합니다.
      parameters:
      - description: RGW Bucket Name
        in: query
        name: bucket_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Delete of RADOS Gateway Bucket
      tags:
      - RGW-Bucket
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: RADOS Gateway 버킷을 리스트를 보여줍니다.
      parameters:
      - description: RGW Bucket Name
        in: query
        name: bucket_name
        type: string
      - default: "false"
        description: RGW Bucket List Detail
        enum:
        - "true"
        - "false"
        in: query
        name: detail
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show List of RADOS Gateway Bucket
      tags:
      - RGW-Bucket
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: RADOS Gateway 버킷을 생성합니다.
      parameters:
      - description: RGW Bucket Name
        in: formData
        name: bucket_name
        required: true
        type: string
      - description: RGW User Name
        in: formData
        name: username
        required: true
        type: string
      - default: false
        description: RGW Bucket Lock Enabled
        enum:
        - true
        - false
        in: formData
        name: lock_enabled
        required: true
        type: boolean
      - description: RGW Bucket Lock Mode
        enum:
        - compliance
        - governance
        in: formData
        name: lock_mode
        type: string
      - description: RGW Bucket Lock Period
        in: formData
        name: lock_retention_period_days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Create of RADOS Gateway Bucket
      tags:
      - RGW-Bucket
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: RADOS Gateway 버킷을 수정합니다.
      parameters:
      - description: RGW Bucket Name
        in: formData
        name: bucket_name
        required: true
        type: string
      - description: RGW Bucket ID
        in: formData
        name: bucket_id
        required: true
        type: string
      - description: RGW User Name
        in: formData
        name: username
        required: true
        type: string
      - description: RGW Bucket Lock Enabled
        enum:
        - Enabled
        - Suspended
        in: formData
        name: versioning
        type: string
      - description: RGW Bucket Lock Mode(Required value if the lock box is checked)
        enum:
        - compliance
        - governance
        in: formData
        name: lock_mode
        type: string
      - description: RGW Bucket Lock Period(Required value if the lock box is checked)
        in: formData
        name: lock_retention_period_days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Update of RADOS Gateway Bucket
      tags:
      - RGW-Bucket
  /api/v1/rgw/quota:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: RADOS Gateway Quota를 설정 및 활성화합니다.
      parameters:
      - description: RGW User ID Name
        in: formData
        name: username
        required: true
        type: string
      - description: RGW Quota Target
        enum:
        - user
        - bucket
        in: formData
        name: scope
        required: true
        type: string
      - description: RGW Quota Max Objects
        in: formData
        name: max_objects
        required: true
        type: integer
      - description: RGW Quota Max Size(B/K/M/G/T)
        in: formData
        name: max_size
        required: true
        type: string
      - default: enable
        description: RGW Quota Whether Activated
        enum:
        - enable
        - disable
        in: formData
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Setting of RADOS Gateway Quota
      tags:
      - RGW
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"